remove unnecessary items
